+++ tests/testthat/test-region_isos.R
@@ -20,3 +20,64 @@ test_that("outputs regions for expected scenario sources", {
     sort(expected_sources)
   )
 })
+
+test_that("outputs expected regions for weo_2020 (#253).", {
+  expected_regions <- c(
+    "advanced economies",
+    "africa",
+    "asia pacific",
+    "caspian",
+    "central and south america",
+    "china",
+    "developing asia",
+    "eurasia",
+    "europe",
+    "european union",
+    "latin america",
+    "middle east",
+    "north africa",
+    "north america",
+    "oecd",
+    "opec",
+    "other african countries and territories",
+    "other asia pacific countries and territories",
+    "other central and south american countries and territories",
+    "southeast asia",
+    "sub saharan africa",
+    "global",
+    "developing economies",
+    "iea",
+    "non oecd",
+    "non opec"
+  )
+
+  region_isos_weo_2020 <- region_isos[region_isos$source == "weo_2020",]
+
+  expect_equal(
+    sort(unique(region_isos_weo_2020$region)),
+    sort(expected_regions)
+  )
+})
+
+test_that("outputs expected regions for isf_2020 (#253).", {
+  expected_regions <- c(
+    "oecd north america",
+    "oecd europe",
+    "global",
+    "latin america",
+    "india",
+    "china",
+    "africa",
+    "middle east",
+    "eurasia/eastern europe",
+    "non oecd asia",
+    "oecd pacific"
+  )
+
+  region_isos_isf_2020 <- region_isos[region_isos$source == "isf_2020",]
+
+  expect_equal(
+    sort(unique(region_isos_isf_2020$region)),
+    sort(expected_regions)
+  )
+})