+++ NEWS.md
@@ -1,5 +1,8 @@
 # r2dii.data (development version)
 
+* `region_isos` now outputs all expected `region`s for: `weo_2020` and 
+  `isf_2020` (#253 @Antoine-Lalechere).
+
 # r2dii.data 0.2.1
 
 * `region_isos` gains data for new scenario `source`s: `weo_2020`, `isf_2020`